Fixing test and improving remove code
