+++ src/main.test.js
@@ -1,7 +1,8 @@
 import * as Yoox from './main'
-import {mixedStore, officeStore, rootAccessors, storeWithModule} from "./test/mockedStore"
+import { mixedStore, officeStore, rootAccessors, storeWithModule } from "./test/mockedStore"
 import Persist from './persist/main'
 import { delay } from "./test/helpers"
+import { NotFoundException } from "./persist/exceptions"
 
 describe('Testing Yoox modularized', () => {
   it('Should create yoox from object', async() => {
@@ -117,5 +118,90 @@ describe('Testing store with more then one module', () => {
       office: { name: 'Cyberlabs' },
       user: { name: 'Gabriel Santiago', age: 25 }
     })
+  })
+})
+
+
+describe('Testing Persist setter', () => {
+  it('Should set data on store when persist is empty', async () => {
+    const cacheResult = await Persist.set({
+      state: {
+        user: {
+          name: 'Gabriel Rizzo',
+          age: 25
+        }
+      }
+    })
+
+    expect(cacheResult).toBeTruthy()
+  })
+
+  it('Should update data on store', async () => {
+    const expectedValue = {
+      state: {
+        group: {
+          id: 1,
+          name: 'Happy Group'
+        }
+      }
+    }
+
+    await Persist.set({
+      state: {
+        user: {
+          name: 'Gabriel Rizzo',
+          age: 25
+        }
+      }
+    })
+
+    await Persist.set(expectedValue)
+
+    const cachedValue = await Persist.get()
+
+    expect(cachedValue).toStrictEqual(expectedValue)
+  })
+})
+
+describe('Testing Persist getter', () => {
+  it('Should get cached value', async () => {
+    const expectedValue = {
+      state: {
+        group: {
+          id: 1,
+          name: 'Happy Group'
+        }
+      }
+    }
+
+    await Persist.set(expectedValue)
+
+    const cachedData = await Persist.get()
+
+    expect(cachedData).toStrictEqual(expectedValue)
+  })
+
+  it('Should return error when there is no cached value', async () => {
+    try {
+      // setting value to guarantee that store is not empty
+      await Persist.set({ value: true })
+
+      // Cleaning persist
+      await Persist.clear()
+
+      await Persist.get()
+    } catch (e) {
+      expect(e instanceof NotFoundException).toBeTruthy()
+    }
+  })
+})
+
+describe('Testing clear persist', () => {
+  it('Should clear stored value', async () => {
+    await Persist.set({ value: true })
+
+    const clearPersistResult = await Persist.clear()
+
+    expect(clearPersistResult).toBeTruthy()
   })
 })
